update(dep): llama.cpp to support longer context
Instead of exposing the customizable rope parameters as options,
the rope frequency scale is calculated from context_token_limit.
This may not always yield the best parameters due to how the model was
trained (e.g., the 8 Ki SuperHOT running with 4 Ki context length),
but it's working well enough for now.